enqueue current in the init command
