+++ bitloops/src/capability_packs/semantic_clones/stage_embeddings_tests.rs
@@ -1,7 +1,9 @@
 use super::*;
+use crate::host::devql::esc_pg;
 use crate::host::devql::{sqlite_exec_path_allow_create, sqlite_query_rows_path};
 use crate::host::inference::{EmbeddingInputType as HostEmbeddingInputType, EmbeddingService};
 use serde_json::json;
+use std::sync::Arc;
 use tempfile::tempdir;
 
 const TEST_EMBEDDINGS_DRIVER: &str = crate::host::inference::BITLOOPS_EMBEDDINGS_IPC_DRIVER;
